+++ src/main/java/com/googlecode/javaewah/BitmapStorage.java
@@ -70,5 +70,5 @@ public interface BitmapStorage {
      *  
      * @param size         the size in bits
      */
-    void setSizeInBitsWithinLastWord(final int size);
+    void setSizeInBitsWithinLastWord(final long size);
 }
